requirements.txt
youtube-transcript-api
groq
streamlit
python-dotenv
